BaddieState.moveDone is s8: retire the BADDIE_MOVE_STATUS_SIGNED escape hatch
Same screen as moveJustStartedA: 103 sites read moveDone through
*(s8*)&/*(char*)& and 44 more through an (s8) cast. Types it s8 unconditionally,
removes 85 puns/casts across 15 DLLs, and deletes the now-dead #define/#undef
sandwich in 437.c, which was the last live user of the macro.

All 991 objects byte-identical; main.dol unchanged.

+++ src/dlls/engine/25/25.c
@@ -793,7 +793,7 @@ int dll_19_shouldDropTarget(GameObject* obj, void* state, f32 distThreshold, int
     GameObject* player = Obj_GetPlayerObject();
     int result = 0;
 
-    if ((s8)((BaddieState*)state)->moveDone != 0)
+    if (((BaddieState*)state)->moveDone != 0)
     {
         if (((BaddieState*)state)->targetObj == player && (s8)((BaddieState*)state)->hitPoints != 0)
         {
